requirements.txt
ebaysdk==2.2.0
google-api-python-client==2.123.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
requests==2.31.0 